feat: add new signup UI
